Precompile curvature folder
